// src/components/Customers/HistoryTable/type.ts
export interface Order {
  id: string;
  date: string;
  customerId: string;
  city: string;
  channel: string;
  status: string;
  total: number;
  items: Array<{
    sku: string;
    name: string;
    qty: number;
    price: number;
  }>;
  comment?: string;
}

export interface OrderHistoryTableProps {
  orders: Order[];
}